% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputies_get_history_of_parliametary_service.R
\name{deputies_get_history_of_services}
\alias{deputies_get_history_of_services}
\title{Get national parties, eu groups, postions}
\usage{
deputies_get_history_of_services(home_page, deputy_id)
}
\description{
Get national parties, eu groups, postions
}
