% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_more_info.R
\name{get_more_info}
\alias{get_more_info}
\title{Get nationality, date of birth, place of birth and/or date of death}
\usage{
get_more_info(home_page)
}
\description{
Get nationality, date of birth, place of birth and/or date of death
}
