% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{read_config}
\alias{read_config}
\title{Loads data from configuration file}
\usage{
read_config(file = "db_config.txt", delim = " ")
}
\arguments{
\item{delim}{a delim parametr in read_delim}

\item{name}{file name or path to file}
}
\value{
A tibble with dbname, host, username and password for database conncetion.
}
\description{
File should contains:
\itemize{
 \item dbname = "dbname"
 \item host = "host"
 \item username = "username"
 \item password = "password"
}
}
\examples{
read_config()
read_config(file = "path/name.txt", delim = " ")

}
