% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statements_core.R
\name{statements_core}
\alias{statements_core}
\title{Get statements text and details.}
\usage{
statements_core(url)
}
\description{
Function \code{statements_core} downloads content and details(language, time) of the statement.
}
