% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statements_get_all_statements.R
\name{statements_get_all_statements}
\alias{statements_get_all_statements}
\title{Get all statements for P8}
\usage{
statements_get_all_statements(deputy_id, browser, term_of_office = 8)
}
\description{
Get all statements for P8
}
