% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statements_get_list_of_statements.R
\name{statements_get_list_of_statements}
\alias{statements_get_list_of_statements}
\title{Return links, titles, dates of sttamenets of given deputie.}
\usage{
statements_get_list_of_statements(deputy_id, browser, term_of_office = 8)
}
\description{
Return links, titles, dates of sttamenets of given deputie.
}
