% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statements_get_statement.R
\name{statements_get_statement}
\alias{statements_get_statement}
\title{Get statements text and details.}
\usage{
statements_get_statement(url)
}
\description{
Function \code{statements_core} downloads content and details(language, time) of the statement.
}
