% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statements_get_time.R
\name{statements_get_time}
\alias{statements_get_time}
\title{Get time of statements}
\usage{
statements_get_time(url)
}
\arguments{
\item{url}{A url of statements}
}
\value{
Returns duration, start time and end time of statement
}
\description{
Get time of statements
}
