% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statements_update_statements.R
\name{statements_update_statements}
\alias{statements_update_statements}
\title{Update stamenets in data abse}
\usage{
statements_update_statements(deputy_id, browser, term_of_office = 8, db)
}
\description{
Update stamenets in data abse
}
