% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_of_statements.R
\name{strings_identical}
\alias{strings_identical}
\title{Remove white signs and comapre two strings}
\usage{
strings_identical(x, y)
}
\value{
Returns TRUE or FALSE
}
\description{
Remove white signs and comapre two strings
}
