% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_of_statements.R
\name{time_of_statements}
\alias{time_of_statements}
\title{Get time of statements}
\usage{
time_of_statements(url)
}
\arguments{
\item{url}{A url of statements}
}
\value{
Returns duration, start time and end time of statement
}
\description{
Get time of statements
}
