% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/europarl.R
\name{europarl}
\alias{europarl}
\alias{europarl_download}
\title{Install Europarl Corpora}
\usage{
europarl_download(verbose = TRUE)
}
\arguments{
\item{verbose}{A \code{logical} value, whether to output progress messages.}
}
\description{
The European Parliament Proceedings Parallel Corpus 1996-2011, known as
[Europarl](http://www.statmt.org/europarl/) is a classical resource in the
CWB community for demonstration purposes. This package includes minimal
functionality to download and install the corpus into the R package.
}
